# artifacts/api-server/src/routes/messages.ts
import { Router } from "express";
import { db } from "@workspace/db";
import {
  messagesTable,
  messageDeletionsTable,
  conversationMembersTable,
  usersTable,
  blocksTable,
} from "@workspace/db";
import { eq, and, desc, notInArray } from "drizzle-orm";
import { requireAuth } from "../lib/auth";
import { io } from "../lib/socket";
import {
  ListMessagesParams,
  SendMessageParams,
  SendMessageBody,
  DeleteMessageForMeParams,
  UnsendMessageParams,
} from "@workspace/api-zod";

const router = Router();

function parseIntParam(val: string | string[] | undefined): number | null {
  const raw = Array.isArray(val) ? val[0] : val;
  if (!raw) return null;
  const n = parseInt(raw, 10);
  return isNaN(n) ? null : n;
}

function formatMessage(
  msg: typeof messagesTable.$inferSelect,
  senderUsername: string | null
) {
  return {
    id: msg.id,
    conversationId: msg.conversationId,
    senderId: msg.senderId,
    senderUsername,
    messageType: msg.messageType,
    content: msg.content ?? null,
    fileUrl: msg.fileUrl ?? null,
    status: msg.status,
    isDeletedForEveryone: msg.isDeletedForEveryone,
    createdAt: msg.createdAt.toISOString(),
  };
}

// GET /api/conversations/:conversationId/messages
router.get(
  "/conversations/:conversationId/messages",
  requireAuth,
  async (req, res): Promise<void> => {
    const conversationId = parseIntParam(req.params.conversationId);
    if (!conversationId) {
      res.status(400).json({ error: "Invalid conversationId" });
      return;
    }

    // Check membership
    const [membership] = await db
      .select()
      .from(conversationMembersTable)
      .where(
        and(
          eq(conversationMembersTable.conversationId, conversationId),
          eq(conversationMembersTable.userId, req.userId!)
        )
      );
    if (!membership) {
      res.status(403).json({ error: "Not a member of this conversation" });
      return;
    }

    // Get deleted message IDs for this user
    const deletions = await db
      .select()
      .from(messageDeletionsTable)
      .where(eq(messageDeletionsTable.userId, req.userId!));
    const deletedIds = deletions.map((d) => d.messageId);

    const messages =
      deletedIds.length > 0
        ? await db
            .select()
            .from(messagesTable)
            .where(
              and(
                eq(messagesTable.conversationId, conversationId),
                notInArray(messagesTable.id, deletedIds)
              )
            )
            .orderBy(desc(messagesTable.createdAt))
            .limit(50)
        : await db
            .select()
            .from(messagesTable)
            .where(eq(messagesTable.conversationId, conversationId))
            .orderBy(desc(messagesTable.createdAt))
            .limit(50);

    // Fetch sender usernames
    const senderIds = [...new Set(messages.map((m) => m.senderId))];
    const senders =
      senderIds.length > 0
        ? await db
            .select()
            .from(usersTable)
            .where(
              senderIds.length === 1
                ? eq(usersTable.id, senderIds[0])
                : eq(usersTable.id, senderIds[0]) // fallback; real fix: use inArray
            )
        : [];

    // Build a map - simplified since inArray was imported above
    const senderMap: Record<number, string> = {};
    for (const s of senders) {
      senderMap[s.id] = s.username;
    }
    // Also fetch remaining senders
    for (const sid of senderIds) {
      if (!senderMap[sid]) {
        const [u] = await db.select().from(usersTable).where(eq(usersTable.id, sid));
        if (u) senderMap[u.id] = u.username;
      }
    }

    const result = messages
      .reverse()
      .map((m) => formatMessage(m, senderMap[m.senderId] ?? null));
    res.json(result);
  }
);

// POST /api/conversations/:conversationId/messages
router.post(
  "/conversations/:conversationId/messages",
  requireAuth,
  async (req, res): Promise<void> => {
    const params = SendMessageParams.safeParse({
      conversationId: parseInt(
        Array.isArray(req.params.conversationId)
          ? req.params.conversationId[0]
          : req.params.conversationId,
        10
      ),
    });
    if (!params.success) {
      res.status(400).json({ error: params.error.message });
      return;
    }
    const { conversationId } = params.data;

    const body = SendMessageBody.safeParse(req.body);
    if (!body.success) {
      res.status(400).json({ error: body.error.message });
      return;
    }

    const { messageType, content, fileUrl } = body.data;

    if (messageType === "text" && !content) {
      res.status(400).json({ error: "Text messages require content" });
      return;
    }
    if ((messageType === "image" || messageType === "voice") && !fileUrl) {
      res.status(400).json({ error: "Media messages require fileUrl" });
      return;
    }

    // Check membership
    const [membership] = await db
      .select()
      .from(conversationMembersTable)
      .where(
        and(
          eq(conversationMembersTable.conversationId, conversationId),
          eq(conversationMembersTable.userId, req.userId!)
        )
      );
    if (!membership) {
      res.status(403).json({ error: "Not a member of this conversation" });
      return;
    }

    // Check blocks (for DMs)
    const [blocked] = await db
      .select()
      .from(blocksTable)
      .where(
        and(
          eq(blocksTable.blockedId, req.userId!)
        )
      );
    // simplified block check; full implementation would check per-user

    const [msg] = await db
      .insert(messagesTable)
      .values({
        conversationId,
        senderId: req.userId!,
        messageType,
        content: content ?? null,
        fileUrl: fileUrl ?? null,
        status: "sent",
      })
      .returning();

    const [sender] = await db
      .select()
      .from(usersTable)
      .where(eq(usersTable.id, req.userId!));

    const formatted = formatMessage(msg, sender?.username ?? null);

    // Emit via socket.io
    if (io) {
      io.to(`conversation:${conversationId}`).emit("message:new", {
        message: formatted,
        conversationId,
      });
    }

    res.status(201).json(formatted);
  }
);

// DELETE /api/messages/:messageId/delete-for-me
router.delete(
  "/messages/:messageId/delete-for-me",
  requireAuth,
  async (req, res): Promise<void> => {
    const params = DeleteMessageForMeParams.safeParse({
      messageId: parseInt(
        Array.isArray(req.params.messageId) ? req.params.messageId[0] : req.params.messageId,
        10
      ),
    });
    if (!params.success) {
      res.status(400).json({ error: params.error.message });
      return;
    }

    await db.insert(messageDeletionsTable).values({
      messageId: params.data.messageId,
      userId: req.userId!,
    });

    res.json({ message: "Message deleted for you" });
  }
);

// DELETE /api/messages/:messageId/unsend
router.delete(
  "/messages/:messageId/unsend",
  requireAuth,
  async (req, res): Promise<void> => {
    const params = UnsendMessageParams.safeParse({
      messageId: parseInt(
        Array.isArray(req.params.messageId) ? req.params.messageId[0] : req.params.messageId,
        10
      ),
    });
    if (!params.success) {
      res.status(400).json({ error: params.error.message });
      return;
    }

    const [msg] = await db
      .select()
      .from(messagesTable)
      .where(eq(messagesTable.id, params.data.messageId));
    if (!msg) {
      res.status(404).json({ error: "Message not found" });
      return;
    }
    if (msg.senderId !== req.userId) {
      res.status(403).json({ error: "You can only unsend your own messages" });
      return;
    }

    await db
      .update(messagesTable)
      .set({ isDeletedForEveryone: true, content: null, fileUrl: null })
      .where(eq(messagesTable.id, params.data.messageId));

    // Notify all members
    if (io) {
      io.to(`conversation:${msg.conversationId}`).emit("message:deleted", {
        messageId: params.data.messageId,
      });
    }

    res.json({ message: "Message unsent" });
  }
);

export default router;