# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Goatup Chat API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users by username
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSummary"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Conversations
...
  /conversations/{conversationId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a conversation (last 50 messages)
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message to a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden (blocked)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /messages/{messageId}/delete-for-me:
    delete:
      operationId: deleteMessageForMe
      tags: [messages]
      summary: Delete a message for the current user only
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /messages/{messageId}/unsend:
    delete:
      operationId: unsendMessage
      tags: [messages]
      summary: Unsend (delete for everyone) - sender only
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unsent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "403":
          description: Not the sender
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Groups
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # Auth schemas
...
    User:
      type: object
      required: [id, username, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        profilePicUrl:
          type: ["string", "null"]
        aboutStatus:
          type: ["string", "null"]
        onlineStatus:
          type: boolean
        lastSeenVisibility:
          type: string
          enum: [everyone, nobody]
        createdAt:
          type: string
    UserSummary:
      type: object
      required: [id, username]
      properties:
        id:
          type: integer
        username:
          type: string
        profilePicUrl:
          type: ["string", "null"]
        aboutStatus:
          type: ["string", "null"]
        onlineStatus:
          type: boolean
    ProfileUpdate:
      type: object
      properties:
        aboutStatus:
          type: ["string", "null"]
        profilePicUrl:
          type: ["string", "null"]
        lastSeenVisibility:
          type: string
          enum: [everyone, nobody]

    # Conversation schemas
...
    Message:
      type: object
      required: [id, conversationId, senderId, messageType, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        senderUsername:
          type: ["string", "null"]
        messageType:
          type: string
          enum: [text, image, voice]
        content:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [sent, delivered, read]
        isDeletedForEveryone:
          type: boolean
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [messageType]
      properties:
        messageType:
          type: string
          enum: [text, image, voice]
        content:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]

    # Group schemas
...