# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Goatup Chat API
 * OpenAPI spec version: 0.1.0
 */
import type { MessageInputMessageType } from './messageInputMessageType';

export interface MessageInput {
  messageType: MessageInputMessageType;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  fileUrl?: string | null;
}